requirements.txt
langchain>=0.1.0
langchain-community>=0.0.16
langchain-openai>=0.0.5
langchain-core>=0.1.4
langchain-vllm>=0.0.1
vllm>=0.3.0
yfinance>=0.2.32
python-dotenv>=1.0.0
pydantic>=2.5.0
numpy>=1.26.0
sentence-transformers>=2.2.2
tiktoken>=0.5.1
colorama>=0.4.6
